requirements.txt
absl-py==0.12.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.21
dgl-cuda10.1==0.6.0.post1
docopt==0.6.2
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gensim==4.0.1
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
graphviz==0.16
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.5.0
ipywidgets==7.6.3
jedi==0.18.0
joblib==1.0.1
jupyterlab-widgets==1.0.0
karateclub==1.2.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.2.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas==1.2.5
plotly==5.0.0
prompt-toolkit==3.0.18
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycombo==0.1.5
pydotz==1.5.1
PyGSP==0.5.1
pylint==2.6.0
python-louvain==0.15
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
sklearn==0.0
smart-open==5.1.0
sympy==1.8
tabulate==0.8.9
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu-estimator==2.3.0
tensorwatch==0.9.1
termcolor==1.1.0
terminado==0.9.4
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm==4.61.0
typed-ast==1.4.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zmq==0.0.0
